Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-values/minmax-percentage-serialize-expected.txt


FAIL e.style['margin-left'] = "min(1%)" should set the property value assert_equals: serialization should be canonical expected "min(1%)" but got "calc(min(1%))"
FAIL e.style['margin-left'] = "max(1%)" should set the property value assert_equals: serialization should be canonical expected "max(1%)" but got "calc(max(1%))"
FAIL e.style['margin-left'] = "min(1%, 2%, 3%)" should set the property value assert_equals: serialization should be canonical expected "min(1%, 2%, 3%)" but got "calc(min(1%, 2%, 3%))"
FAIL e.style['margin-left'] = "min(3%, 2%, 1%)" should set the property value assert_equals: serialization should be canonical expected "min(3%, 2%, 1%)" but got "calc(min(3%, 2%, 1%))"
FAIL e.style['margin-left'] = "max(1%, 2%, 3%)" should set the property value assert_equals: serialization should be canonical expected "max(1%, 2%, 3%)" but got "calc(max(1%, 2%, 3%))"
FAIL e.style['margin-left'] = "max(3%, 2%, 1%)" should set the property value assert_equals: serialization should be canonical expected "max(3%, 2%, 1%)" but got "calc(max(3%, 2%, 1%))"
FAIL e.style['margin-left'] = "calc(min(1%) + min(2%))" should set the property value assert_equals: serialization should be canonical expected "calc(min(1%) + min(2%))" but got "calc((min(1%)) + (min(2%)))"
FAIL e.style['margin-left'] = "calc(max(1%) + max(2%))" should set the property value assert_equals: serialization should be canonical expected "calc(max(1%) + max(2%))" but got "calc((max(1%)) + (max(2%)))"
FAIL e.style['margin-left'] = "calc(1% + min(1%))" should set the property value assert_equals: serialization should be canonical expected "calc(1% + min(1%))" but got "calc(1% + (min(1%)))"
FAIL e.style['margin-left'] = "calc(min(1%) + 1%)" should set the property value assert_equals: serialization should be canonical expected "calc(1% + min(1%))" but got "calc((min(1%)) + 1%)"
FAIL e.style['margin-left'] = "calc(1% + max(1%))" should set the property value assert_equals: serialization should be canonical expected "calc(1% + max(1%))" but got "calc(1% + (max(1%)))"
FAIL e.style['margin-left'] = "calc(max(1%) + 1%)" should set the property value assert_equals: serialization should be canonical expected "calc(1% + max(1%))" but got "calc((max(1%)) + 1%)"
